requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu126

torch==2.6.*
torchvision==0.21.*
transformers>=4.51.0
accelerate>=0.34.0
tokenizers>=0.20.0
safetensors>=0.4.5
huggingface-hub>=0.25.0
pillow>=10.3.0